pyproject.toml
[tool.poetry]
name = "binance-connect"
version = "0.1.0"
description = ""
authors = ["Даниил Павленко <d.pavlenko@mixvel.com>"]
readme = "README.md"
packages = [{include = "binance_connect"}]

[tool.poetry.dependencies]
python = "^3.10"
certifi = "^2022.12.7"
python-dotenv = "^0.21.1"
pytest = "^7.2.1"
pandas = "^1.5.3"
binance-futures-connector = "^3.3.0"
gspread = "^5.7.2"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
